fix: Resolve circular dependencies by deferred access and a new module
